perf: precompute PSF FFT and regularization matrices for speed
- Precompute PSF FFT once in PixelizedLensSimulator to avoid repeated FFT calculations
- Precompute unit regularization matrices to eliminate redundant scaling operations per half_size
- Change data loading and model image dtype to float32 for reduced memory usage

+++ TinyLensGpu/utils/misc.py
@@ -29,9 +29,9 @@ def load_lens_data(
     mask : np.ndarray or None
         Boolean mask (True = masked out) if provided
     """
-    image_data = fits.getdata(image_path).astype("float64")
-    noise_map = fits.getdata(noise_path).astype("float64")
-    psf_kernel = fits.getdata(psf_path).astype("float64")
+    image_data = fits.getdata(image_path).astype("float32")
+    noise_map = fits.getdata(noise_path).astype("float32")
+    psf_kernel = fits.getdata(psf_path).astype("float32")
 
     mask = None
     if mask_path is not None:
